Renamed package for the IME scanning classes.
