+++ package/tecla/keyboard/TeclaIME.java
@@ -1,4 +1,4 @@
-package ca.idrc.tecla.framework;
+package com.android.tecla.keyboard;
 
 import android.inputmethodservice.InputMethodService;
 import android.inputmethodservice.Keyboard;
